refactor(eval): simplify cache and analysis helpers
