+++ evalscope/evaluator/evaluator.py
@@ -20,10 +20,10 @@ from evalscope.api.judge import summarize_judge_runs
 from evalscope.api.metric import AggScore, SampleScore
 from evalscope.api.registry import register_evaluator
 from evalscope.constants import HEARTBEAT_INTERVAL_SEC, ScoreStatus
-from evalscope.evaluation_versioning import ResolvedBenchmarkSpec, build_analysis_context, build_benchmark_identity
+from evalscope.evaluation_versioning import ResolvedBenchmarkSpec, build_benchmark_identity
 from evalscope.evaluator.batch_reviewer import BatchReviewer
 from evalscope.evaluator.perf_collector import PerfCollector
-from evalscope.report import Report, gen_perf_table, gen_table
+from evalscope.report import Report, build_analysis_context, gen_perf_table, gen_table
 from evalscope.utils.function_utils import run_in_threads_with_progress
 from evalscope.utils.logger import get_logger
