+++ evalscope/report/__init__.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from .generator import ReportGenerator
     from .ref import ReportRef
     from .renderer import gen_html_report_file
-    from .report import Category, Metric, Report, ReportKey, Subset
+    from .report import BenchmarkAnalysisContext, Category, Metric, Report, ReportKey, Subset, build_analysis_context
 
 else:
     _import_structure = {
@@ -39,10 +39,12 @@ else:
         ],
         'report': [
             'Category',
+            'BenchmarkAnalysisContext',
             'Report',
             'ReportKey',
             'Subset',
             'Metric',
+            'build_analysis_context',
         ],
         'renderer': [
             'gen_html_report_file',
